requirements.txt
Flask
flask-cors
pdfminer.six
python-docx
scikit-learn
